pyproject.toml
[project]
name = "version_2_adk_agent"
version = "0.1.0"
description = "A2A Agent With Google ADK"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.1.8",
    "fastapi>=0.115.12",
    "google-adk>=0.2.0",
    "google-genai>=1.11.0",
    "python-dotenv>=1.1.0",
]

[tool.setuptools]
packages = ["app", "agents", "server", "models", "client"]
